# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  handler.CreateFollowRecordRequest:
    properties:
      appointmentTime:
        type: string
      content:
        type: string
      customerId:
        type: integer
      customerLevelId:
        type: integer
      customerSourceId:
        type: integer
      followMethodId:
        type: integer
      nextFollowTime:
        type: string
      shootingTime:
        type: string
    required:
    - content
    - customerId
    type: object
...
  model.DashboardMonthlyContractCount:
    properties:
      count:
        type: integer
      label:
        type: string
    type: object
  model.DashboardMonthlyRevenue:
    properties:
      amount:
        type: number
      label:
        type: string
    type: object
  model.DashboardOverview:
    properties:
      conversionRate:
        $ref: '#/definitions/model.DashboardStat'
      monthlyContracts:
        items:
          $ref: '#/definitions/model.DashboardMonthlyContractCount'
        type: array
      monthlyRevenue:
        items:
          $ref: '#/definitions/model.DashboardMonthlyRevenue'
        type: array
      newCustomers:
        $ref: '#/definitions/model.DashboardStat'
      newOpportunities:
        $ref: '#/definitions/model.DashboardStat'
      recentActivities:
        items:
          $ref: '#/definitions/model.DashboardRecentActivity'
        type: array
      recentDeals:
        items:
          $ref: '#/definitions/model.DashboardRecentDeal'
        type: array
      revenue:
        $ref: '#/definitions/model.DashboardStat'
    type: object
  model.DashboardRecentActivity:
    properties:
      action:
        type: string
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      target:
        type: string
      type:
        type: string
      userName:
        type: string
    type: object
  model.DashboardRecentDeal:
    properties:
      amount:
        type: number
      contractName:
        type: string
      createdAt:
        type: string
      customerEmail:
        type: string
      customerName:
        type: string
      id:
        type: integer
      userName:
        type: string
    type: object
  model.DashboardStat:
    properties:
      changeRate:
        type: number
      current:
        type: number
      previous:
        type: number
    type: object
...
  model.FollowMethod:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      sort:
        type: integer
    type: object
  model.FollowMethodRequest:
    properties:
      name:
        type: string
      sort:
        type: integer
    required:
    - name
    type: object
  model.OperationFollowRecord:
    properties:
      appointmentTime:
        type: string
      content:
        type: string
      createdAt:
        type: string
      customer:
        $ref: '#/definitions/model.OperationFollowRecordCustomer'
      customerId:
        type: integer
      customerLevelId:
        type: integer
      customerLevelName:
        type: string
      customerSourceId:
        type: integer
      customerSourceName:
        type: string
      followMethodId:
        type: integer
      followMethodName:
        type: string
      id:
        type: integer
      nextFollowTime:
        type: string
      operatorUserId:
        type: integer
      operatorUserName:
        type: string
      shootingTime:
        type: string
      updatedAt:
        type: string
    type: object
  model.OperationFollowRecordCustomer:
    properties:
      area:
        type: integer
      city:
        type: integer
      collectTime:
        type: string
      contactName:
        type: string
      dealStatus:
        type: string
      detailAddress:
        type: string
      email:
        type: string
      followTime:
        type: string
      id:
        type: integer
      legalName:
        type: string
      levelId:
        type: integer
      levelName:
        type: string
      name:
        type: string
      nextTime:
        type: string
      ownerUserId:
        type: integer
      ownerUserName:
        type: string
      primaryPhone:
        type: string
      province:
        type: integer
      remark:
        type: string
      sourceId:
        type: integer
      sourceName:
        type: string
      status:
        type: string
      weixin:
        type: string
    type: object
  model.OperationFollowRecordListResult:
    properties:
      items:
        items:
          $ref: '#/definitions/model.OperationFollowRecord'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  model.SalesFollowRecord:
    properties:
      content:
        type: string
      createdAt:
        type: string
      customer:
        $ref: '#/definitions/model.OperationFollowRecordCustomer'
      customerId:
        type: integer
      customerLevelId:
        type: integer
      customerLevelName:
        type: string
      customerSourceId:
        type: integer
      customerSourceName:
        type: string
      followMethodId:
        type: integer
      followMethodName:
        type: string
      id:
        type: integer
      nextFollowTime:
        type: string
      operatorUserId:
        type: integer
      operatorUserName:
        type: string
      updatedAt:
        type: string
    type: object
  model.SalesFollowRecordListResult:
    properties:
      items:
        items:
          $ref: '#/definitions/model.SalesFollowRecord'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  service.CustomerAutoDropTaskFailure:
    properties:
      customerId:
        type: integer
      error:
        type: string
    type: object
  service.CustomerAutoDropTaskResult:
    properties:
      bothRulesMatchedDropped:
        type: integer
      dealDropDays:
        type: integer
      dealTimeoutDropped:
        type: integer
      dropped:
        type: integer
      evaluated:
        type: integer
      executedAt:
        type: string
      failures:
        items:
          $ref: '#/definitions/service.CustomerAutoDropTaskFailure'
        type: array
      followUpDropDays:
        type: integer
      followUpTimeoutDropped:
        type: integer
      holidayModeEnabled:
        type: boolean
      skipReason:
        type: string
      skipped:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: 后端服务API文档
  title: Backend API
  version: "1.0"
paths:
...
  /api/v1/dashboard/overview:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.DashboardOverview'
              type: object
        "401":
          description: 未登录或登录已失效
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: 获取仪表盘概览
      tags:
      - dashboard
...
  /api/v1/follow-methods:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.FollowMethod'
                  type: array
              type: object
      summary: 获取所有跟进方式
      tags:
      - 跟进记录
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建跟进方式
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.FollowMethodRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.APIResponse'
      summary: 创建跟进方式
      tags:
      - 跟进记录
  /api/v1/follow-methods/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 跟进方式ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.APIResponse'
      summary: 删除跟进方式
      tags:
      - 跟进记录
    put:
      consumes:
      - application/json
      parameters:
      - description: 跟进方式ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新跟进方式
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.FollowMethodRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.APIResponse'
      summary: 更新跟进方式
      tags:
      - 跟进记录
...
  /api/v1/operation-follow-records:
    get:
      consumes:
      - application/json
      parameters:
      - description: 客户ID
        format: int64
        in: query
        name: customerId
        required: true
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.OperationFollowRecordListResult'
              type: object
      summary: 获取运营跟进记录列表
      tags:
      - 跟进记录
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建运营跟进记录
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateFollowRecordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.APIResponse'
      summary: 创建运营跟进记录
      tags:
      - 跟进记录
  /api/v1/operation-follow-records/all:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.OperationFollowRecordListResult'
              type: object
      summary: 获取所有运营跟进记录列表
      tags:
      - 跟进记录
...
  /api/v1/sales-follow-records:
    get:
      consumes:
      - application/json
      parameters:
      - description: 客户ID
        format: int64
        in: query
        name: customerId
        required: true
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.SalesFollowRecordListResult'
              type: object
      summary: 获取销售跟进记录列表
      tags:
      - 跟进记录
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建销售跟进记录
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateFollowRecordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.APIResponse'
      summary: 创建销售跟进记录
      tags:
      - 跟进记录
  /api/v1/sales-follow-records/all:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.SalesFollowRecordListResult'
              type: object
      summary: 获取所有销售跟进记录列表
      tags:
      - 跟进记录
...
  /api/v1/tasks/customer-drop/run:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/service.CustomerAutoDropTaskResult'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: 执行客户自动掉库任务
      tags:
      - crontab
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token, 格式: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"